tests

+++ backend/controllers/user/deleteUser.ts
@@ -1,9 +1,12 @@
-import { PrismaClient } from '../../generated/prisma';
+import prisma from '../../prisma/client';
 import { Request, Response } from 'express';
 import asyncHandler from 'express-async-handler';
-const prisma = new PrismaClient();
+const { BetterError } = require('../../middleware/errorHandler');
 const deleteUser = asyncHandler(async (req: Request, res: Response) => {
   const id = Number(req.params.id);
+  if (!id) {
+    throw new BetterError('Wrong id', 404, 'WRONG_ID', 'User Error');
+  }
   try {
     await prisma.users.delete({
       where: {
@@ -14,7 +17,12 @@ const deleteUser = asyncHandler(async (req: Request, res: Response) => {
       message: 'deleted user successfully',
     });
   } catch (err) {
-    throw new Error('no user found');
+    throw new BetterError(
+      'User not deleted',
+      404,
+      'USER_NOT_DELETED',
+      'User Error'
+    );
   }
 });
